Refactor helper class to take into account view lifecycle
There are now 2 constructors of this class. User can choose which one suits them. Details are on the comments.

The original code has a bug that is when a view (Fragment or Activity) is destroyed after its onViewCreated invoked and this class constructor is called onViewCreated, the callback that might has some UI updates will be called after onDestroy invoked, which is an illegal action. For example, root view will be null at the time of callback invoked and when you query viewLifecycleOwner somewhere in that callback, it will fail.

The 4 arguments constructor has similar behavior with the original constructor. Except, in this one, we observe the lifecycle of the view passed on to this constructor. While this works, it only works only if the constructor called on onCreate or after that(?).